re-implemented next into clothingitem.js

+++ controllers/clothingItem.js
@@ -24,32 +24,28 @@ const createItem = (req, res, next) => {
     .then((item) => {
       res.send({ data: item });
     })
+    // .catch((err) => {
+    //   handleError(err, res);
+    // });
     .catch((err) => {
-      handleError(err, res);
+      if (err.name === "ValidationError") {
+        next(new BadRequestError("Bad request, invalid data"));
+      } else {
+        next(err);
+      }
     });
-  // .catch(() => {
-  //   next(
-  //     new BadRequestError(
-  //       "The request to create an item is invalid. Please, try again."
-  //     )
-  //   );
-  // });
 };
 
 const getItems = (req, res, next) => {
   ClothingItem.find()
     .sort({ createdAt: -1 })
     .then((items) => res.status(200).send({ data: items }))
+    // .catch((err) => {
+    //   handleError(err, res);
+    // });
     .catch((err) => {
-      handleError(err, res);
+      next(err);
     });
-  // .catch(() => {
-  //   next(
-  //     new NotFoundError(
-  //       "The requested items cannot be found. Please, try again."
-  //     )
-  //   );
-  // });
 };
 
 const updateItem = (req, res, next) => {
@@ -57,12 +53,16 @@ const updateItem = (req, res, next) => {
   const { imageUrl } = req.body;
 
   ClothingItem.findOneAndUpdate(itemId, { $set: { imageUrl } })
-    .orFail(() => {
-      handleOnFailError();
+    .then((item) => {
+      if (!item) {
+        next(new NotFoundError("Item not found"));
+        return;
+      } else {
+        res.status(200).send({ data: item });
+      }
     })
-    .then((item) => res.status(200).send({ data: item }))
     .catch((err) => {
-      handleError(err, res);
+      next(err);
     });
   // .catch(() => {
   //   next(
@@ -73,34 +73,48 @@ const updateItem = (req, res, next) => {
   // });
 };
 
+// const updateItem = (req, res, next) => {
+//   const { itemId } = req.param;
+//   const { imageUrl } = req.body;
+
+//   ClothingItem.findOneAndUpdate(itemId, { $set: { imageUrl } })
+//     .orFail(() => {
+//       handleOnFailError();
+//     })
+//     .then((item) => res.status(200).send({ data: item }))
+//     .catch((err) => {
+//       handleError(err, res);
+//     });
+// .catch(() => {
+//   next(
+//     new UnauthorizedError(
+//       "You are not authorized to update this item. Please, update items that you have created."
+//     )
+//   );
+// });
+// };
+
 const deleteItem = (req, res, next) => {
   ClothingItem.findById(req.params.itemId)
-    .orFail(() => {
-      const error = new Error("Item ID not found");
-      error.statusCode = 404;
-      throw error;
-    })
     .then((item) => {
-      if (String(item.owner) !== req.user._id) {
-        return res
-          .status(ERROR_CODES.Forbidden)
-          .send({ message: "You are not authorized to delete this item" });
+      if (!item) {
+        next(new NotFoundError("Item not found"));
+        return;
       }
-      return item.deleteOne().then(() => {
-        res.send({ message: "Item deleted" });
-      });
+      if (String(item.owner) !== req.user._id) {
+        next(new ForbiddenError("You are not authorized to delete this item"));
+        return;
+      } else {
+        return item.deleteOne().then(() => {
+          res.send({ message: "Item deleted" });
+        });
+      }
     })
     .catch((err) => {
-      if (err.statusCode === 404) {
-        res.status(ERROR_CODES.NotFound).send({ message: "Item not found" });
-      } else if (err.name === "CastError") {
-        res
-          .status(ERROR_CODES.BadRequest)
-          .send({ message: "Bad Request and/or invalid input" });
+      if (err.name === "CastError") {
+        next(new BadRequestError("Invalid item ID"));
       } else {
-        res
-          .status(ERROR_CODES.DefaultError)
-          .send({ message: "Something went wrong" });
+        next(err);
       }
     });
   // .catch(() => {
@@ -108,25 +122,85 @@ const deleteItem = (req, res, next) => {
   // });
 };
 
+// const deleteItem = (req, res, next) => {
+//   ClothingItem.findById(req.params.itemId)
+//     .orFail(() => {
+//       const error = new Error("Item ID not found");
+//       error.statusCode = 404;
+//       throw error;
+//     })
+//     .then((item) => {
+//       if (String(item.owner) !== req.user._id) {
+//         return res
+//           .status(ERROR_CODES.Forbidden)
+//           .send({ message: "You are not authorized to delete this item" });
+//       }
+//       return item.deleteOne().then(() => {
+//         res.send({ message: "Item deleted" });
+//       });
+//     })
+//     .catch((err) => {
+//       if (err.statusCode === 404) {
+//         res.status(ERROR_CODES.NotFound).send({ message: "Item not found" });
+//       } else if (err.name === "CastError") {
+//         res
+//           .status(ERROR_CODES.BadRequest)
+//           .send({ message: "Bad Request and/or invalid input" });
+//       } else {
+//         res
+//           .status(ERROR_CODES.DefaultError)
+//           .send({ message: "Something went wrong" });
+//       }
+//     });
+// .catch(() => {
+//   next();
+// });
+// };
+
 const likeItem = (req, res, next) => {
   ClothingItem.findByIdAndUpdate(
     req.params.itemId,
     { $addToSet: { likes: req.user._id } }, // add _id to the array if it's not there yet
     { new: true }
   )
-    .orFail(() => {
-      handleOnFailError();
-    })
     .then((item) => {
-      res.status(200).send({ data: { ...item.toObject() } });
+      if (!item) {
+        next(new NotFoundError("Item not found"));
+      } else {
+        res.status(200).send({ data: { ...item.toObject() } });
+      }
     })
     .catch((err) => {
-      handleError(err, res);
+      if (err.name === "CastError") {
+        next(new BadRequestError("Bad request, invalid data ID"));
+      } else {
+        next(err);
+      }
     });
   // .catch(() => {
   //   next(new ConflictError("Something went wrong. Please, try again later."));
   // });
 };
+
+// const likeItem = (req, res, next) => {
+//   ClothingItem.findByIdAndUpdate(
+//     req.params.itemId,
+//     { $addToSet: { likes: req.user._id } }, // add _id to the array if it's not there yet
+//     { new: true }
+//   )
+//     .orFail(() => {
+//       handleOnFailError();
+//     })
+//     .then((item) => {
+//       res.status(200).send({ data: { ...item.toObject() } });
+//     })
+//     .catch((err) => {
+//       handleError(err, res);
+//     });
+// .catch(() => {
+//   next(new ConflictError("Something went wrong. Please, try again later."));
+// });
+// };
 
 const dislikeItem = (req, res, next) =>
   ClothingItem.findByIdAndUpdate(
@@ -134,14 +208,19 @@ const dislikeItem = (req, res, next) =>
     { $pull: { likes: req.user._id } }, // remove _id from the array
     { new: true }
   )
-    .orFail(() => {
-      handleOnFailError();
-    })
     .then((item) => {
-      res.status(200).send({ data: { ...item.toObject() } });
+      if (!item) {
+        next(new NotFoundError("Item not found"));
+      } else {
+        res.status(200).send({ data: { ...item.toObject() } });
+      }
     })
     .catch((err) => {
-      handleError(err, res);
+      if (err.name === "CastError") {
+        next(new BadRequestError("Bad request, invalid data ID"));
+      } else {
+        next(err);
+      }
     });
 // .catch(() => {
 //   next(new ConflictError("Something went wrong. Please, try again later."));
